+++ src/Synthesizer/AGenerator.cpp
@@ -1,6 +1,6 @@
 #include "AGenerator.h"
 
-void synthesizer::AGenerator::generate(noteBuffer& noteBuffer, const uchar* keyState, const settings& settings, const double* dynamicsProfile, const double* releaseProfile){
+void synthesizer::AGenerator::generate(noteBuffer& noteBuffer, const uchar* keyState, const settings& settings, const float* dynamicsProfile, const float* releaseProfile){
     uint i = 0;
     for (; i < settings.sampleSize; i++){
         if (keyState[i]){
@@ -10,7 +10,7 @@ void synthesizer::AGenerator::generate(noteBuffer& noteBuffer, const uchar* keyS
             noteBuffer.releaseSamplesPassed = 0;
             noteBuffer.velocity = keyState[i]/127.0;
 
-            double dynamicsMultiplier = settings.dynamicsDuration > noteBuffer.pressSamplessPassed ? dynamicsProfile[noteBuffer.pressSamplessPassed] : settings.fadeTo;
+            float dynamicsMultiplier = settings.dynamicsDuration > noteBuffer.pressSamplessPassed ? dynamicsProfile[noteBuffer.pressSamplessPassed] : settings.fadeTo;
 
             noteBuffer.buffer[i] = soundFunction(noteBuffer) * settings.volume * noteBuffer.velocity * dynamicsMultiplier;
 
@@ -18,7 +18,7 @@ void synthesizer::AGenerator::generate(noteBuffer& noteBuffer, const uchar* keyS
             noteBuffer.phaze++;
             noteBuffer.releaseSamplesPassed++;
 
-            double dynamicsMultiplier = settings.dynamicsDuration > noteBuffer.pressSamplessPassedCopy ? dynamicsProfile[noteBuffer.pressSamplessPassedCopy] : settings.fadeTo;
+            float dynamicsMultiplier = settings.dynamicsDuration > noteBuffer.pressSamplessPassedCopy ? dynamicsProfile[noteBuffer.pressSamplessPassedCopy] : settings.fadeTo;
 
 
             noteBuffer.pressSamplessPassed = (settings.attack.duration - 1) * dynamicsMultiplier;
