+++ src/Synthesizer/DynamicsController.cpp
@@ -7,13 +7,13 @@ void DynamicsController::calculateDynamicsProfile(settings& settings){
     noteDynamicsProfile.clear();
 
     for (uint i = 0; i < settings.attack.duration; i++){
-        noteDynamicsProfile.push_back(i/double(settings.attack.duration));
+        noteDynamicsProfile.push_back(i/float(settings.attack.duration));
     }
     for (uint i = 0; i < settings.sustain.duration; i++){
         noteDynamicsProfile.push_back(1);
     }
     for (uint i = 0; i < settings.fade.duration; i++){
-        noteDynamicsProfile.push_back((1-i/double(settings.fade.duration)) * (1-settings.fadeTo) + settings.fadeTo);
+        noteDynamicsProfile.push_back((1-i/float(settings.fade.duration)) * (1-settings.fadeTo) + settings.fadeTo);
     }
 
     settings.dynamicsDuration = noteDynamicsProfile.size();
@@ -28,7 +28,7 @@ void DynamicsController::calculateReleaseProfile(settings& settings){
     noteReleaseProfile.clear();
 
     for (uint i = 0; i < settings.release.duration; i++){
-        noteReleaseProfile.push_back(1-i/double(settings.release.duration));
+        noteReleaseProfile.push_back(1-i/float(settings.release.duration));
     }
 }
 
@@ -41,8 +41,8 @@ void DynamicsController::calculateReleaseProfile(settings& settings){
 //     const ulong& size = noteReleaseProfile.size();
 //     releaseDynamicMap = new uint[size];
 //     for (uint i = 0, j = noteDynamicsProfile.size() - 1; i < size; i++){
-//         double lastDifference;
-//         double difference = std::abs(noteDynamicsProfile.at(j) - noteReleaseProfile.at(i));
+//         float lastDifference;
+//         float difference = std::abs(noteDynamicsProfile.at(j) - noteReleaseProfile.at(i));
 //         do {
 //             lastDifference = difference;
 //             j--;
@@ -52,11 +52,11 @@ void DynamicsController::calculateReleaseProfile(settings& settings){
 //     }
 // }
 
-const double* DynamicsController::getDynamicsProfile(){
+const float* DynamicsController::getDynamicsProfile(){
     return noteDynamicsProfile.data();
 }
 
-const double* DynamicsController::getReleaseProfile(){
+const float* DynamicsController::getReleaseProfile(){
     return noteReleaseProfile.data();
 }
