+++ src/Synthesizer/Generator_Noise1.cpp
@@ -1,5 +1,5 @@
 #include "Generator_Noise1.h"
 
-inline double synthesizer::Generator_Noise1::soundFunction(noteBuffer& noteBuffer){
+inline float synthesizer::Generator_Noise1::soundFunction(noteBuffer& noteBuffer){
     return sin(noteBuffer.phaze*noteBuffer.multiplier);
 }
